pyproject.toml
[project]
name = "another-s3-manager"
version = "0.1.0"
description = "Lightweight web-based file manager for Amazon S3 buckets built with FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "boto3==1.29.7",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.1",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "pytest-mock==3.12.0",
    "httpx==0.27.0",
    "moto[s3]==5.1.2",
]

[tool.setuptools]
# Explicitly specify Python modules (not packages) to avoid auto-discovery issues
py-modules = ["auth", "config", "constants", "main", "s3_client", "users", "utils"]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
